feat(nix)!: Bundle dark-notify binary with Vim plugin
This patch adds the binary as a `runtimeDeps` of the Vim plugin. Now
downstream consumers don't have to worry about properly installing
dark-notify alongside the plugin.

This does of course mean that if one had installed a modified binary
globally, this would no longer be used by the binary. So this is
technically a breaking change.

BREAKING CHANGE: Plugin will ignore globally installed binary.